implemented first index check
