+++ mofchecker/checks/local_structure/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .overlapping_atoms import AtomicOverlapCheck
