+++ mofchecker/checks/oms/errors.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+from ...errors import NoMetal
+
+
+class LowCoordinationNumber(KeyError):
+    """Error for low coordination number"""
+
+
+class HighCoordinationNumber(KeyError):
+    """Error for high coordination number"""
+
+
+class NoOpenDefined(KeyError):
+    """Error in case the open check is not defined
+    for this coordination number"""
