+++ mofchecker/errors.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+class NoMetal(KeyError):
+    """Error in case there is no metal in structure"""
