+++ tests/checks/global_structure.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from mofchecker.checks.global_structure import (
+    HasCarbon,
+    HasHydrogen,
+    HasMetal,
+    HasNitrogen,
+)
+
+
+def test_no_c(get_no_c):
+    for structure in get_no_c:
+        has_c = HasCarbon(structure)
+        assert not has_c.is_ok
+
+
+def test_no_h(get_no_h):
+    for structure in get_no_h:
+        has_h = HasHydrogen(structure)
+        assert not has_h.is_ok
